+++ BasicRenderer/include/RenderPasses/LightCullingPass.h
@@ -69,7 +69,7 @@ public:
 		commandList->SetComputeRoot32BitConstants(LightClusterRootSignatureIndex, NumLightClusterRootConstants, lightClusterConstants, 0);
 
 		auto clusterSize = getClusterSize();
-		unsigned int numThreadGroups = std::ceil(((float)(clusterSize.x * clusterSize.y * clusterSize.z)) / 128);
+		unsigned int numThreadGroups = static_cast<unsigned int>(std::ceil(((float)(clusterSize.x * clusterSize.y * clusterSize.z)) / 128));
 		commandList->Dispatch(numThreadGroups, 1, 1);
 		return {};
 	}
