+++ BasicRenderer/include/RenderPasses/PostProcessing/SpecularIBLPass.h
@@ -65,7 +65,7 @@ public:
         CD3DX12_CPU_DESCRIPTOR_HANDLE rtvHandle(m_pHDRTarget->GetRTVInfo(0).cpuHandle);
         commandList->OMSetRenderTargets(1, &rtvHandle, FALSE, nullptr);
 
-        CD3DX12_VIEWPORT viewport = CD3DX12_VIEWPORT(0.0f, 0.0f, context.renderResolution.x, context.renderResolution.y);
+        CD3DX12_VIEWPORT viewport = CD3DX12_VIEWPORT(0.0f, 0.0f, static_cast<float>(context.renderResolution.x), static_cast<float>(context.renderResolution.y));
         CD3DX12_RECT scissorRect = CD3DX12_RECT(0, 0, context.renderResolution.x, context.renderResolution.y);
         commandList->RSSetViewports(1, &viewport);
         commandList->RSSetScissorRects(1, &scissorRect);
